Split util::encoding
