+++ libmamba/tests/CMakeLists.txt
@@ -25,6 +25,7 @@ set(
     src/util/test_cast.cpp
     src/util/test_compare.cpp
     src/util/test_cryptography.cpp
+    src/util/test_encoding.cpp
     src/util/test_environment.cpp
     src/util/test_flat_bool_expr_tree.cpp
     src/util/test_flat_set.cpp
